Nearly done synchronous question asking system.
Any Employee can ask a question to their supervisor, and the question
is escalated upwards 50% of the time. The way to ask a question is to
schedule the askQuestion Runnable of the Employee class. When this
Runnable gets its turn, it will pause execution of Runnables from its
queue until the question is answered, at which point things resume. A
bit more work is needed for resuming, but I should be able to get it
done today. Feel free to work on things now, but stay out of Scheduler
and Employee ;)